be quiet by default
don't output debugging information by default, that way the output of
this tool can be piped around without stripping out extra information
first.

to do this I had to add the ability to process arguments.  to do THAT I
added underscore to ratiocinate.js, so the 'require' line will need to
be upgraded when underscore is updated/moved.  There might be a more
desirable way to do this.

as a final note, the verbose option is brittle because `verbose` needs
to be considered when logging information out (are we logging output or
messages?).  I was going to fix this by unifying logging under a
'lib/log.js' file, or by overriding `console.log`, but this change
seemed to be the smallest, so I opted for it instead :)